Initial JSON configuration parsing code and package layout
Using JSON to specify most of the configurable server parameters. Command line arguments are limited to a path to a config file containing that JSON and a DNSSEC key used to authenticate update and zone transfer requests throughout a rendezvous cluster. Began laying out stub files for remaining functional piece.