Need to rework implementation somewhat after issues in testing
The ISC dhcpd server cannot be configured with multiple primary DNS servers, so the main record update/notification implementation flow for Hive is not possible in the dhcpd/BIND paired environment. Need to add support for NOTIFY events from a DNS master, IXFRs to incrementally update the Hive instance cache of the local zone, and clarify the exact data transmission role of the Hive instance to its peers.

In the process, I've tweaked the WriteUpdate implementation to properly transform records for deletion. I think there's still more refactoring to be done there.